Fix delete item feature test so it clicks submit button
Javascript in the browser adds an onclick event to the delete form.
This code doesn't appear to be running.  So, the
browser.click('.delete-form') command doesn't work.  Adding the
browser.execute function causes the browser to submit the form as
though it were clicked.

Removed second assert looking for the image source URL because there
are no items on the page.  So, there are no "item-img" classes.